requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
playwright==1.41.0
openai==1.10.0
beautifulsoup4==4.12.3
pandas==2.2.0
python-multipart==0.0.6
python-dotenv==1.0.1
matplotlib==3.8.2
scipy==1.12.0
pypdf==4.0.1
openpyxl==3.1.5
lxml==5.1.0
openai-whisper==20231117
word2number==1.1
annotated-types==0.7.0
pydantic==2.12.4
pydantic-core==2.41.5
typing-extensions==4.15.0